Move presenter extensions
